entity: adjust route entity data type

+++ src/main/java/restful/api/eztrain/entity/RouteEntity.java
@@ -45,7 +45,7 @@ public class RouteEntity {
     private Double tripDistance;
 
     @Column(nullable = false)
-    private Integer tripDuration;
+    private Double tripDuration;
 
     @CreationTimestamp
     @Column(updatable = false, name = "created_at")
